requirements.txt
fastapi==0.111.1
sqlalchemy==2.0.31
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic-settings==2.3.4
aioredis==2.0.1
async-timeout==4.0.3